+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.3.15] - 2026-01-11
 
+### Changed
+
+- Reduced the size of binary files by optimizing lingua-go import. (#450)
+
 ### Fixed
 
 - Fixed an issue where the old database can not be migrated correctly in some cases.